Split out 'Flight Structure' into it's own files
