+++ server/events/index.js
@@ -4,7 +4,11 @@ import "./engines";
 import "./sensors";
 import "./transporters";
 import "./clients";
-import "./flightStructure";
+import "./flight";
+import "./simulator";
+import "./stationSet";
+import "./timeline";
+import "./mission";
 import "./shipStructure";
 import "./shields";
 import "./lrComm";
